Rendering_service/report_agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
httpx==0.27.2
python-multipart==0.0.12
pydantic==2.10.4
pillow==10.1.0
python-dotenv==1.0.0
